Add displayName and username properties to post data
Previously, the post data only included the author uid. To display the author's name and username on the blog post, additional HTTP requests were made to retrieve this information. This approach resulted in slower performance and increased resource usage.

To improve performance, this commit adds the displayName and username properties to the post data. These properties are retrieved from the server when the post is created, eliminating the need for additional HTTP requests. This approach reduces the size of the response and improves the overall performance of the website